Create MajorContourForLargeText.py
This will help in detecting from an image which are the Text or Item in an image which are greater than normal text. This will help in separating Handwritten part, signature part and Logos or Stamps from an image. Also here we are storing the Co-ordinates of those contours so as we can crop images according to those contours and co-ordinates from an original image.