Create HistogramMachineLearningModel.py
Try to take avg. histogram value from a test image and Create a model for histogram from sample images. Then Try that histogram value to match other images. If those images lies within the range of histogram than there will be a possibility that, that part of image is a signature.
(try to take sample test image which is confirmed that its a signature and try to create a Histogram model from it. Histogram helps us to detect the closest possibility of similar image shape)